requirements.txt
aiosqlite==0.17.0
alembic==1.6.2
appdirs==1.4.4
asyncpg==0.22.0
black==21.5b1
certifi==2020.12.5
click==7.1.2
cloudinary==1.25.0
colorama==0.4.4
fastapi==0.65.0
greenlet==1.1.0
h11==0.12.0
importlib-metadata==4.0.1
Mako==1.1.4
MarkupSafe==2.0.0
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
pydantic==1.8.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-multipart==0.0.5
PyYAML==5.4.1
regex==2021.4.4
six==1.16.0
SQLAlchemy==1.4.15
starlette==0.14.2
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
zipp==3.4.1
